server: move tcpKeepAliveManager out of server.go
Release note: None
