+++ pkg/server/BUILD.bazel
@@ -46,6 +46,7 @@ go_library(
         "statements.go",
         "status.go",
         "sticky_engine.go",
+        "tcp_keepalive_manager.go",
         "tenant.go",
         "tenant_status.go",
         "testing_knobs.go",
